+++ tests/testthat/test-file_coverage.R
@@ -1,5 +1,7 @@
 context("file_coverage")
-writeLines(con = "s1",
+s1 <- tempfile()
+t1 <- tempfile()
+writeLines(con = s1,
 "a <- function(x) {
   x + 1
 }
@@ -9,7 +11,7 @@ b <- function(x) {
   else FALSE
 }")
 
-writeLines(con = "t1",
+writeLines(con = t1,
 "a(1)
 a(2)
 a(3)
@@ -17,10 +19,10 @@ b(0)
 b(1)
 b(2)")
 
-on.exit(unlink(c("s1", "t1")))
+on.exit(unlink(c(s1, t1)))
 
 test_that("it works on single files", {
-  cov <- file_coverage("s1", "t1")
+  cov <- file_coverage(s1, t1)
   cov_d <- as.data.frame(cov)
 
   expect_equal(cov_d$functions, c("a", "b", "b", "b"))